swagger 테스트 및 JWT 작성중,,, 팀장님돌아와

+++ moit-v3/back/src/main/java/com/moit/member/dto/AuthUserDto.java
@@ -12,8 +12,8 @@ public class AuthUserDto {
 	private String nickname;
 	private String typeName;
 	private String profileUrl;
-	private Integer statusId;
-	private Integer memberTypeId;
+	private Long statusId;
+	private Long memberTypeId;
 
 	private  List<AuthDto> authList;
